Compass: Consistently use (const) qreal

+++ src/sensors/compass.h
@@ -23,6 +23,7 @@
 
 #include <memory>
 
+#include <QtGlobal>
 #include <QObject>
 
 class QMetaMethod;
@@ -54,9 +55,9 @@ public:
 	
 	/** Returns the most recent azimuth value
 	 *  (in degrees clockwise from north; updated approx. every 30 milliseconds). */
-	float getCurrentAzimuth();
+	qreal getCurrentAzimuth();
 	
-	/** Connects to the azimuthChanged(float azimuth_degrees) signal. This ensures to use a queued
+	/** Connects to the azimuthChanged(qreal azimuth_degrees) signal. This ensures to use a queued
 	 *  connection, which is important because the data provider runs on another
 	 *  thread. Updates are delivered approx. every 30 milliseconds. */
 	void connectToAzimuthChanges(const QObject* receiver, const char* slot);
@@ -67,7 +68,7 @@ public:
 signals:
 	/** Emitted regularly with the current azimuth value (in degrees).
 	 *  Preferably use connectToAzimuthChanges() to connect to this signal. */
-	void azimuthChanged(float azimuth);
+	void azimuthChanged(qreal azimuth);
 	
 protected:
 	void connectNotify(const QMetaMethod& signal) override;
@@ -75,7 +76,7 @@ protected:
 	void disconnectNotify(const QMetaMethod& signal) override;
 	
 private:
-	void emitAzimuthChanged(float value);
+	void emitAzimuthChanged(qreal value);
 	
 	std::unique_ptr<CompassPrivate> p;
 	int reference_counter = 0;
